+++ tests/integration/test_use_cases/test_04_agent_registration.py
@@ -16,7 +16,7 @@ def test_04_agent_register_and_resolve(network, config, accounts):
 
     remote_agent = config.agent_list['remote']
     ddo = RemoteAgent.generate_ddo(remote_agent['url'])
-    authentication_access = {
+    authentication = {
         'username': remote_agent['username'],
         'password': remote_agent['password']
     }
@@ -24,13 +24,13 @@ def test_04_agent_register_and_resolve(network, config, accounts):
     register_account = accounts[0]
 
     did = ddo.did
-    remote_agent = RemoteAgent.register(network, register_account, ddo, authentication_access)
+    remote_agent = RemoteAgent.register(network, register_account, ddo, authentication)
     assert(remote_agent)
     found_ddo = DDO(json_text=network.resolve_did(did))
     assert(found_ddo.as_text() == ddo.as_text())
 
 
-    resolved_agent = RemoteAgent.load(network, ddo.did, authentication_access=authentication_access)
+    resolved_agent = RemoteAgent.load(network, ddo.did, authentication=authentication)
     assert(resolved_agent)
     assert(resolved_agent.ddo)
     assert(resolved_agent.ddo.as_text() == ddo.as_text())